requirements.txt
torchsummary
numpy
Pillow
opencv-python
yaml
scipy
tqdm
lpips